@@ -1,49 +1,65 @@
 {
   "name": "pi-client",
-  "version": "0.0.8",
+  "version": "dev",
+  "private": true,
   "scripts": {
     "ng": "ng",
-    "start": "ng serve",
-    "build": "ng build",
+    "start": "npx ts-node ./script/prebuild.ts -e development && ng serve --configuration development",
+    "build:dev": "npx ts-node ./script/prebuild.ts -e development && ng build --configuration development",
+    "build:prod": "npx ts-node ./script/prebuild.ts -e production && ng build --configuration production",
     "test": "ng test",
-    "lint": "ng lint",
-    "e2e": "ng e2e"
+    "lint:fix": "pretty-quick --staged && ng lint --fix",
+    "e2e": "ng e2e",
+    "prepare": "cd ../ && husky install piClient/.husky"
   },
-  "private": true,
   "dependencies": {
-    "@angular/animations": "~11.0.4",
-    "@angular/common": "~11.0.4",
-    "@angular/compiler": "~11.0.4",
-    "@angular/core": "~11.0.4",
-    "@angular/forms": "~11.0.4",
-    "@angular/platform-browser": "~11.0.4",
-    "@angular/platform-browser-dynamic": "~11.0.4",
-    "@angular/router": "~11.0.4",
-    "chart.js": "^2.9.4",
+    "@angular/animations": "~14.0.1",
+    "@angular/common": "~14.0.1",
+    "@angular/compiler": "~14.0.1",
+    "@angular/core": "~14.0.1",
+    "@angular/forms": "~14.0.1",
+    "@angular/platform-browser": "~14.0.1",
+    "@angular/platform-browser-dynamic": "~14.0.1",
+    "@angular/router": "~14.0.1",
+    "chart.js": "^3.8.0",
+    "eslint-plugin-import": "^2.26.0",
+    "eslint-plugin-jsdoc": "^39.3.2",
+    "eslint-plugin-prefer-arrow": "^1.2.3",
     "material-design-icons": "^3.0.1",
-    "ng2-charts": "^2.4.2",
-    "rxjs": "~6.6.0",
-    "tslib": "^2.0.0",
-    "zone.js": "~0.10.2"
+    "ng2-charts": "^3.0.11",
+    "rxjs": "~7.5.5",
+    "tslib": "^2.4.0",
+    "zone.js": "~0.11.5"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.1100.4",
-    "@angular/cli": "~11.0.4",
-    "@angular/compiler-cli": "~11.0.4",
-    "@ngx-env/builder": "^1.1.0",
-    "@types/jasmine": "~3.6.0",
-    "@types/node": "^12.11.1",
-    "codelyzer": "^6.0.0",
-    "jasmine-core": "~3.6.0",
-    "jasmine-spec-reporter": "~5.0.0",
-    "karma": "~6.3.16",
-    "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage": "~2.0.3",
-    "karma-jasmine": "~4.0.0",
-    "karma-jasmine-html-reporter": "^1.5.0",
+    "@angular-devkit/build-angular": "~14.0.1",
+    "@angular-eslint/builder": "14.0.0-alpha.0",
+    "@angular-eslint/eslint-plugin": "14.0.0-alpha.0",
+    "@angular-eslint/eslint-plugin-template": "14.0.0-alpha.0",
+    "@angular-eslint/schematics": "14.0.0-alpha.0",
+    "@angular-eslint/template-parser": "14.0.0-alpha.0",
+    "@angular/cli": "~14.0.1",
+    "@angular/compiler-cli": "~14.0.1",
+    "@ngx-env/builder": "^2.2.0",
+    "@types/jasmine": "~4.0.3",
+    "@types/node": "^17.0.41",
+    "@typescript-eslint/eslint-plugin": "5.17.0",
+    "@typescript-eslint/parser": "5.17.0",
+    "eslint": "^8.12.0",
+    "husky": "^8.0.0",
+    "jasmine-core": "~4.1.1",
+    "jasmine-spec-reporter": "~7.0.0",
+    "karma": "~6.3.20",
+    "karma-chrome-launcher": "~3.1.1",
+    "karma-coverage": "~2.2.0",
+    "karma-jasmine": "~5.0.1",
+    "karma-jasmine-html-reporter": "^2.0.0",
+    "npm-run-all": "^4.1.5",
+    "prettier": "^2.6.2",
+    "pretty-quick": "^3.1.3",
     "protractor": "~7.0.0",
-    "ts-node": "~8.3.0",
-    "tslint": "~6.1.0",
-    "typescript": "~4.0.2"
+    "ts-node": "^10.8.1",
+    "typescript": "~4.7.3",
+    "yargs": "^17.5.1"
   }
 }
