@@ -7,7 +7,7 @@ import { NavigationComponent } from './navigation/navigation.component';
 import { HttpClientModule } from '@angular/common/http';
 import { AppRoutingModule } from './app-routing.module';
 import { WeatherDefaultComponent } from './weather/weather-default/weather-default.component';
-import { ChartsModule } from 'ng2-charts';
+import { NgChartsModule } from 'ng2-charts';
 import { WeatherSevenDayChartComponent } from './weather/weather-seven-day-chart/weather-seven-day-chart.component';
 import { WeatherHourlyChartComponent } from './weather/weather-hourly-chart/weather-hourly-chart.component';
 
@@ -18,15 +18,10 @@ import { WeatherHourlyChartComponent } from './weather/weather-hourly-chart/weat
     NavigationComponent,
     WeatherDefaultComponent,
     WeatherSevenDayChartComponent,
-    WeatherHourlyChartComponent
+    WeatherHourlyChartComponent,
   ],
-  imports: [
-    BrowserModule,
-    HttpClientModule,
-    AppRoutingModule,
-    ChartsModule
-  ],
+  imports: [BrowserModule, HttpClientModule, AppRoutingModule, NgChartsModule],
   providers: [],
-  bootstrap: [AppComponent]
+  bootstrap: [AppComponent],
 })
-export class AppModule { }
+export class AppModule {}
