@@ -1,14 +1,13 @@
 import { Component, OnInit } from '@angular/core';
-import { version } from '../../../package.json';
+import packageInfo from '../../../package.json';
 
 @Component({
   selector: 'app-navigation',
   templateUrl: './navigation.component.html',
-  styleUrls: ['./navigation.component.css']
+  styleUrls: ['./navigation.component.css'],
 })
 export class NavigationComponent implements OnInit {
-
-  version: string = version;
+  version: string = packageInfo.version;
   today: number = Date.now();
 
   constructor() {}
