@@ -1,21 +1,17 @@
-import {Component, EventEmitter, Input, OnChanges, OnInit, Output, ViewChild} from '@angular/core';
-import {ChartData, ChartDataSets, ChartLineOptions, ChartOptions, ChartType} from 'chart.js';
-import { Color, BaseChartDirective, Label } from 'ng2-charts';
-import {Coord} from '../../model/weather/coord';
-import {DailyWeatherResponse} from '../../model/weather/daily-weather-response';
-import {WeatherService} from '../weather.service';
-import {finalize} from 'rxjs/operators';
-import {DatePipe} from '@angular/common';
-import {WeatherResponse} from '../../model/weather/weather-response';
-import {ResponseStatus} from '../../model/response-status.enum';
+import { Component, EventEmitter, Input, OnInit, Output, ViewChild } from '@angular/core';
+import { ChartConfiguration, ChartDataset, ChartOptions, ChartType } from 'chart.js';
+import { BaseChartDirective } from 'ng2-charts';
+import { Coord } from '../../model/weather/coord';
+import { DailyWeatherResponse } from '../../model/weather/daily-weather-response';
+import { WeatherService } from '../weather.service';
+import { DatePipe } from '@angular/common';
 
 @Component({
   selector: 'app-weather-seven-day-chart',
   templateUrl: './weather-seven-day-chart.component.html',
-  styleUrls: ['./weather-seven-day-chart.component.css']
+  styleUrls: ['./weather-seven-day-chart.component.css'],
 })
 export class WeatherSevenDayChartComponent implements OnInit {
-
   @Input() city: string;
   @Input() coord: Coord;
   @Input() dailyWeatherResponse: DailyWeatherResponse;
@@ -24,107 +20,77 @@ export class WeatherSevenDayChartComponent implements OnInit {
   dailyWeatherResponseLoading = true;
   errorMessage: string;
 
-  constructor(
-    private weatherService: WeatherService
-  ) { }
+  constructor(private weatherService: WeatherService) {}
 
-  public lineChartData: ChartDataSets[] = [];
-  public lineChartLabels: Label[];
-  public lineChartOptions: (ChartOptions) = {
+  public lineChartData: ChartConfiguration['data'] = { datasets: [], labels: [] };
+  public lineChartOptions: ChartOptions = {
+    elements: {
+      line: {
+        tension: 0.3,
+      },
+    },
     maintainAspectRatio: false,
     responsive: true,
-    legend: {
-      labels: { fontColor: 'white' }
-    },
     scales: {
-      xAxes: [{
-        ticks: { fontColor: 'white', fontSize: 15 },
-        gridLines: { color: 'rgba(255,255,255,0.1)' }
-      }],
-      yAxes: [
-        {
-          id: 'y-axis-0',
-          position: 'left',
-          ticks: {
-            // There is no precision attribute in ng2-charts but in Chart.js. So ignore the warning.
-            // @ts-ignore
-            precision: 0,
-            beginAtZero: true,
-            fontColor: 'white',
-            fontSize: 15,
-            callback: (value, index, values) => {
-              return value + '°';
-            }
-          },
-          gridLines: { color: 'rgba(255,255,255,0.1)' }
+      x: {
+        offset: false,
+        ticks: { color: 'white', font: { size: 15 } },
+        grid: { color: 'rgba(255,255,255,0.1)' },
+      },
+      'y-axis-0': {
+        position: 'left',
+        beginAtZero: true,
+        ticks: {
+          precision: 0,
+          color: 'white',
+          font: { size: 15 },
+          callback: (value, index, values) => value + '°',
         },
-        {
-          id: 'y-axis-1',
-          position: 'right',
-          ticks: {
-            max: 1,
-            stepSize: 0.2,
-            beginAtZero: true,
-            fontColor: 'white',
-            fontSize: 10,
-            callback: (value, index, values) => {
-              return ((value as number) * 100) + '%';
-            }
-          },
+        grid: { color: 'rgba(255,255,255,0.1)' },
+      },
+      'y-axis-1': {
+        position: 'right',
+        max: 1,
+        beginAtZero: true,
+        ticks: {
+          stepSize: 0.2,
+          color: 'white',
+          font: { size: 10 },
+          callback: (value, index, values) => (value as number) * 100 + '%',
         },
-      ]
+      },
     },
     layout: {
       padding: {
         left: 0,
         right: 0,
         top: 0,
-        bottom: 30
-      }
+        bottom: 30,
+      },
+    },
+    plugins: {
+      legend: {
+        labels: { color: 'white' },
+      },
     },
   };
 
   images: string[];
-  public lineChartPlugins = [{
-    afterDraw: chart => {
-      const ctx = chart.chart.ctx;
-      const xAxis = chart.scales['x-axis-0'];
-      const yAxis = chart.scales['y-axis-0'];
-      xAxis.ticks.forEach((value, index) => {
-        const x = xAxis.getPixelForTick(index);
-        const image = new Image();
-        image.src = this.images[index],
-          ctx.drawImage(image, x - 22.5, yAxis.bottom + 25, 45, 45);
-      });
-    }
-  }];
-
-  public lineChartColors: Color[] = [
+  public lineChartPlugins = [
     {
-      backgroundColor: 'rgba(63,165,255,0.2)',
-      borderColor: 'rgba(63,165,255,1)',
-      pointBackgroundColor: 'rgba(63,165,255,1)',
-      pointBorderColor: '#fff',
-      pointHoverBackgroundColor: '#fff',
-      pointHoverBorderColor: 'rgba(63,165,255,1)'
-    },
-    {
-      backgroundColor: 'rgba(255,182,0,0.2)',
-      borderColor: 'rgb(255,182,0)',
-      pointBackgroundColor: 'rgba(255,182,0,1)',
-      pointBorderColor: '#fff',
-      pointHoverBackgroundColor: '#fff',
-      pointHoverBorderColor: 'rgba(255,182,0,1)'
-    },
-    {
-      backgroundColor: 'rgba(205,229,255,0.6)',
-      borderColor: 'rgb(205,229,255)',
-      pointBackgroundColor: 'rgba(205,229,255)',
-      pointBorderColor: '#fff',
-      pointHoverBackgroundColor: '#fff',
-      pointHoverBorderColor: 'rgba(205,229,255)'
+      afterDraw: (chart) => {
+        const ctx = chart.ctx;
+        const xAxis = chart.scales['x'];
+        const yAxis = chart.scales['y-axis-0'];
+        xAxis.ticks.forEach((value, index) => {
+          const x = xAxis.getPixelForTick(index);
+          const image = new Image();
+          (image.src = this.images[index]), ctx.drawImage(image, x - 22.5, yAxis.bottom + 25, 45, 45);
+        });
+      },
     },
   ];
+
   public lineChartLegend = true;
   public lineChartType: ChartType = 'line';
 
@@ -144,8 +110,8 @@ export class WeatherSevenDayChartComponent implements OnInit {
 
   updateDailyWeather(coord: Coord): void {
     const dailyWeatherResponse$ = this.weatherService.getDailyWeatherResponseByCoord(coord);
-    dailyWeatherResponse$
-      .subscribe(r => {
+    dailyWeatherResponse$.subscribe(
+      (r) => {
         if (r.responseStatus.toString() !== 'SUCCESS') {
           this.dailyWeatherResponseLoading = true;
           this.errorMessage = r.responseMessage;
@@ -156,36 +122,70 @@ export class WeatherSevenDayChartComponent implements OnInit {
         this.prepareChart(r);
         this.dailyWeatherResponseLoading = false;
         this.errorMessage = undefined;
-      }, error => {
+      },
+      (error) => {
         this.dailyWeatherResponseLoading = true;
         this.errorMessage = error.message;
         setTimeout(this.updateDailyWeather.bind(this), 6000);
-      });
+      }
+    );
   }
 
   prepareChart(dailyWeatherResponse: DailyWeatherResponse): void {
-    const tempMax: ChartDataSets = { data: [], label: 'max', fill: 'origin' };
-    const tempMin: ChartDataSets = { data: [], label: 'min', fill: 'origin' };
-    const pop: ChartDataSets = { data: [], label: 'Probability of precipitation', type: 'bar', yAxisID: 'y-axis-1' };
-    tempMax.data = dailyWeatherResponse.dailyWeather.weathers.map(weather => {
-      return parseFloat((weather.tempMax - 273.15).toFixed(1));
-    });
-    tempMin.data = dailyWeatherResponse.dailyWeather.weathers.map(weather => {
-      return parseFloat((weather.tempMin - 273.15).toFixed(1));
-    });
-    pop.data = dailyWeatherResponse.dailyWeather.weathers.map(weather => {
-      return weather.pop;
-    });
-    this.lineChartLabels = dailyWeatherResponse.dailyWeather.weathers.map((weather, index) => {
+    const tempMax: ChartDataset = {
+      data: [],
+      label: 'max',
+      fill: 'origin',
+      backgroundColor: 'rgba(63,165,255,0.2)',
+      borderColor: 'rgba(63,165,255,1)',
+      hoverBackgroundColor: '#fff',
+      hoverBorderColor: 'rgb(63,165,255,1)',
+      pointBackgroundColor: 'rgba(63,165,255,1)',
+      pointBorderColor: '#fff',
+      pointHoverBackgroundColor: '#fff',
+      pointHoverBorderColor: 'rgba(63,165,255,1)',
+    };
+    const tempMin: ChartDataset = {
+      data: [],
+      label: 'min',
+      fill: 'origin',
+      backgroundColor: 'rgba(255,182,0,0.2)',
+      borderColor: 'rgb(255,182,0)',
+      hoverBackgroundColor: '#fff',
+      hoverBorderColor: 'rgb(255,182,0)',
+      pointBackgroundColor: 'rgb(255,182,0)',
+      pointBorderColor: '#fff',
+      pointHoverBackgroundColor: '#fff',
+      pointHoverBorderColor: 'rgb(255,182,0)',
+    };
+    const pop: ChartDataset = {
+      data: [],
+      label: 'Probability of precipitation',
+      type: 'bar',
+      yAxisID: 'y-axis-1',
+      backgroundColor: 'rgba(205,229,255,0.6)',
+      borderColor: 'rgb(205,229,255)',
+      hoverBackgroundColor: '#fff',
+      hoverBorderColor: 'rgba(205,229,255)',
+    };
+    tempMax.data = dailyWeatherResponse.dailyWeather.weathers.map((weather) =>
+      parseFloat((weather.tempMax - 273.15).toFixed(1))
+    );
+    tempMin.data = dailyWeatherResponse.dailyWeather.weathers.map((weather) =>
+      parseFloat((weather.tempMin - 273.15).toFixed(1))
+    );
+    pop.data = dailyWeatherResponse.dailyWeather.weathers.map((weather) => weather.pop);
+    const lineChartLabels = dailyWeatherResponse.dailyWeather.weathers.map((weather, index) => {
       if (index === 0) {
         return 'Today';
       }
-      return new DatePipe('en-US').transform(weather.dt * 1000 , 'E');
+      return new DatePipe('en-US').transform(weather.dt * 1000, 'E');
     });
 
-    this.images = dailyWeatherResponse.dailyWeather.weathers.map(weather => {
-      return 'https://openweathermap.org/img/wn/' + weather.icon + '.png';
-    });
-    this.lineChartData.push(tempMin, tempMax, pop);
+    this.images = dailyWeatherResponse.dailyWeather.weathers.map(
+      (weather) => 'https://openweathermap.org/img/wn/' + weather.icon + '.png'
+    );
+    this.lineChartData.datasets.push(tempMin, tempMax, pop);
+    this.lineChartData.labels.push(...lineChartLabels);
   }
 }
